feat/add types to the checkAuth action&reducer
